feat(web): mount TanStack Router and remove the hash hook
Step 2 of the route-migration plan in
docs/experiments/2026-05-04-web-path-forward.md.

- main.tsx mounts <RouterProvider/> instead of <App/>; the root route's
  component is attached at startup to keep lib/router.ts free of route-
  component imports (the route components reference route ids back, so a
  direct import would create a cycle).
- A new RootRoute owns the boot lifecycle, app shell, and global host
  components. Two effect-only siblings live inside it:
  * UrlToStoreHydrator runs in useLayoutEffect to write the matched
    route's params/search into the existing Zustand fields, so
    MainContent's store-driven shape keeps working until step 4 drains
    those fields.
  * StoreToRouterBridge mirrors store changes back into the router via
    useNavigate, only when the URL doesn't already match. It skips work
    on URL-driven renders to avoid racing with the hydrator's
    layout-effect write of a deep-link.
- useHashRouter and the legacyHash phase-0 helpers are deleted along
  with App.tsx; they were only consumed by the deleted hook.
- /console and /threads legacy aliases drop out of router.ts and
  routeRegistry.ts per the plan's open question #8: only the canonical
  /apps/console and /apps/threads remain.

Verified: bunx tsc --noEmit clean, bun run build, 773 web unit tests
green, web/e2e/run-local.sh shell green (31 tests).
